moved dice loss functions into loss_functions.py
